Move prepareCodeTour to separate file
